pyproject.toml
[tool.poetry]
name = "week14"
version = "0.1.0"
description = ""
authors = ["Tomek H. <tomaszheese@o2.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.2.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"
pytest = "^8.0.2"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
